# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart-Browser Bookmarks API",
    "description": "Tag-first bookmarks. Contract version 2. See doc/architecture.md.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/bookmarks/{save_id}/tags": {
      "post": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Add Tags",
        "operationId": "add_tags_api_v2_bookmarks__save_id__tags_post",
        "parameters": [
          {
            "name": "save_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Save Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{save_id}/tags/{tag_name}": {
      "delete": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Remove Tag",
        "operationId": "remove_tag_api_v2_bookmarks__save_id__tags__tag_name__delete",
        "parameters": [
          {
            "name": "save_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Save Id"
            }
          },
          {
            "name": "tag_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/tags": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "List Tags",
        "description": "Your tags, with how often you have used them.\n\nThe vocabulary is global (ADR 0002) but the counts are yours: what belongs in an\nautocomplete list is how often *you* reach for a tag. Ranking by usage rather than\nalphabetically is what makes a 567-entry vocabulary usable at all.",
        "operationId": "list_tags_api_v2_tags_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on tag name",
              "title": "Q"
            },
            "description": "Substring match on tag name"
          },
          {
            "name": "min_count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Min Count"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagOut"
                  },
                  "title": "Response List Tags Api V2 Tags Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BookmarkOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "saved_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved From"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "site",
          "saved_from",
          "created_at",
          "tags"
        ],
        "title": "BookmarkOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TagOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "count"
        ],
        "title": "TagOut"
      },
      "TagsIn": {
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Tags"
          },
          "source": {
            "type": "string",
            "enum": [
              "user",
              "ai",
              "rule",
              "import"
            ],
            "title": "Source",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "tags"
        ],
        "title": "TagsIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}